Rename file to remove "benchmark" from the name.
